Packages refactoring and added unit tests
